lstring2 library tests moved to separate files; added string [] and at()
