+++ tests/src/lvstring2test.h
@@ -1,0 +1,12 @@
+#ifndef LVSTRING2TEST_H
+#define LVSTRING2TEST_H
+
+#include "lvstring2.h"
+
+namespace lv {
+
+void test_lstring2();
+
+}
+
+#endif // LVSTRING2TEST_H
